refactor: BREAKING CHANGE merge with audiopc_ffi
