+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cmake.sourceDirectory": "E:/Documents/Github/audiopc_ffi/example/linux/flutter"
+  "cmake.sourceDirectory": "E:/Documents/Github/audiopc/example/linux/flutter"
 }
