+++ CONTRIBUTE.md
@@ -1,0 +1,4 @@
+# Contributor
+
+* Feel free to contribute to repository
+* Contact me for more
