+++ README.md
@@ -1,4 +1,4 @@
-# audiopc_ffi
+# audiopc
 
 A new Flutter FFI plugin project.
 
@@ -62,8 +62,8 @@ The native build systems that are invoked by FFI (and method channel) plugins ar
 * For Android: Gradle, which invokes the Android NDK for native builds.
   * See the documentation in android/build.gradle.
 * For iOS and MacOS: Xcode, via CocoaPods.
-  * See the documentation in ios/audiopc_ffi.podspec.
-  * See the documentation in macos/audiopc_ffi.podspec.
+  * See the documentation in ios/audiopc.podspec.
+  * See the documentation in macos/audiopc.podspec.
 * For Linux and Windows: CMake.
   * See the documentation in linux/CMakeLists.txt.
   * See the documentation in windows/CMakeLists.txt.
@@ -72,17 +72,17 @@ The native build systems that are invoked by FFI (and method channel) plugins ar
 
 To use the native code, bindings in Dart are needed.
 To avoid writing these by hand, they are generated from the header file
-(`src/audiopc_ffi.h`) by `package:ffigen`.
+(`src/audiopc.h`) by `package:ffigen`.
 Regenerate the bindings by running `dart run ffigen --config ffigen.yaml`.
 
 ## Invoking native code
 
 Very short-running native functions can be directly invoked from any isolate.
-For example, see `sum` in `lib/audiopc_ffi.dart`.
+For example, see `sum` in `lib/audiopc.dart`.
 
 Longer-running functions should be invoked on a helper isolate to avoid
 dropping frames in Flutter applications.
-For example, see `sumAsync` in `lib/audiopc_ffi.dart`.
+For example, see `sumAsync` in `lib/audiopc.dart`.
 
 ## Flutter help
