+++ android/build.gradle
@@ -1,6 +1,6 @@
 // The Android Gradle Plugin builds the native code with the Android NDK.
 
-group = "com.thugbn.audiopc_ffi"
+group = "com.thugbn.audiopc"
 version = "1.0"
 
 buildscript {
@@ -25,7 +25,7 @@ rootProject.allprojects {
 apply plugin: "com.android.library"
 
 android {
-    namespace = "com.thugbn.audiopc_ffi"
+    namespace = "com.thugbn.audiopc"
 
     // Bumping the plugin compileSdk version requires all clients of this plugin
     // to bump the version in their app.
