+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'audiopc_ffi'
+rootProject.name = 'audiopc'
