+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
-name: audiopc_ffi_example
-description: "Demonstrates how to use the audiopc_ffi plugin."
+name: audiopc_example
+description: "Demonstrates how to use the audiopc plugin."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -31,9 +31,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  audiopc_ffi:
+  audiopc:
     # When depending on this package from a real application you should use:
-    #   audiopc_ffi: ^x.y.z
+    #   audiopc: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
