+++ example/windows/flutter/generated_plugins.cmake
@@ -6,7 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
-  audiopc_ffi
+  audiopc
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)
