+++ example/windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "audiopc_ffi_example" "\0"
+            VALUE "FileDescription", "audiopc_example" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "audiopc_ffi_example" "\0"
+            VALUE "InternalName", "audiopc_example" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2026 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "audiopc_ffi_example.exe" "\0"
-            VALUE "ProductName", "audiopc_ffi_example" "\0"
+            VALUE "OriginalFilename", "audiopc_example.exe" "\0"
+            VALUE "ProductName", "audiopc_example" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
