+++ ios/Classes/audiopc_ffi.c
@@ -1,3 +1,3 @@
 // Relative import to be able to reuse the C sources.
-// See the comment in ../audiopc_ffi.podspec for more information.
-#include "../../src/audiopc_ffi.c"
+// See the comment in ../audiopc.podspec for more information.
+#include "../../src/audiopc.c"
