+++ lib/audiopc.dart
@@ -1,2 +1,2 @@
 export 'package:audiopc_interface/audiopc_interface.dart';
-export 'src/audiopc_native.dart';
+export 'src/audio_player.dart';
