+++ macos/audiopc_ffi.podspec
@@ -1,9 +1,9 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint audiopc_ffi.podspec` to validate before publishing.
+# Run `pod lib lint audiopc.podspec` to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'audiopc_ffi'
+  s.name             = 'audiopc'
   s.version          = '0.0.1'
   s.summary          = 'A new Flutter FFI plugin project.'
   s.description      = <<-DESC
@@ -24,7 +24,7 @@ A new Flutter FFI plugin project.
   # data, update the PrivacyInfo.xcprivacy file to describe your plugin's
   # privacy impact, and then uncomment this line. For more information,
   # see https://developer.apple.com/documentation/bundleresources/privacy_manifest_files
-  # s.resource_bundles = {'audiopc_ffi_privacy' => ['Resources/PrivacyInfo.xcprivacy']}
+  # s.resource_bundles = {'audiopc_privacy' => ['Resources/PrivacyInfo.xcprivacy']}
 
   s.dependency 'FlutterMacOS'
