--- src/audiopc_ffi.c
@@ -1,2 +1,0 @@
-#include "audiopc_ffi.h"
-
